More bugfixes? More bugs?

+++ Simulations/RigidBodySystemSimulator.h
@@ -20,7 +20,7 @@ struct Rigidbody
 	Vec3 Torque;
 	Vec3 Force;
 	Mat4 InertiaTensorInv;
-	int Mass;
+	float Mass;
 };
 
 class RigidBodySystemSimulator:public Simulator{
@@ -45,7 +45,7 @@ public:
 	Vec3 getLinearVelocityOfRigidBody(int i);
 	Vec3 getAngularVelocityOfRigidBody(int i);
 	void applyForceOnBody(int i, Vec3 loc, Vec3 force);
-	void addRigidBody(Vec3 position, Vec3 size, int mass);
+	void addRigidBody(Vec3 position, Vec3 size, float mass);
 	void setOrientationOf(int i,Quat orientation);
 	void setVelocityOf(int i, Vec3 velocity);
